move all icon related code to separate helper class
