+++ Classes/ViewHelpers/Be/FunctionBarViewHelper.php
@@ -25,9 +25,7 @@ namespace Slub\SlubEvents\ViewHelpers\Be;
  ***************************************************************/
 
 use TYPO3\CMS\Backend\Utility\BackendUtility;
-use TYPO3\CMS\Backend\Utility\IconUtility;
-use TYPO3\CMS\Core\Imaging\Icon;
-use TYPO3\CMS\Core\Imaging\IconFactory;
+
 use TYPO3\CMS\Extbase\Configuration\ConfigurationManagerInterface;
 use TYPO3\CMS\Extbase\Object\ObjectManagerInterface;
 use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
@@ -40,10 +38,6 @@ class FunctionBarViewHelper extends \TYPO3\CMS\Fluid\ViewHelpers\Be\AbstractBack
      */
     protected $configurationManager;
 
-    /**
-     * @var IconFactory
-     */
-    protected $iconFactory;
 
     /**
      * @var ObjectManagerInterface
@@ -62,13 +56,6 @@ class FunctionBarViewHelper extends \TYPO3\CMS\Fluid\ViewHelpers\Be\AbstractBack
 
         $this->objectManager = $objectManager;
 
-        // iconFactory exists from TYPO3 7
-        if (version_compare(TYPO3_version, '7.6.0', '>=')) {
-
-            $this->iconFactory = $this->objectManager->get('TYPO3\\CMS\\Core\\Imaging\\IconFactory');
-
-        }
-
     }
 
     /**
@@ -79,95 +66,6 @@ class FunctionBarViewHelper extends \TYPO3\CMS\Fluid\ViewHelpers\Be\AbstractBack
     public function injectConfigurationManager(ConfigurationManagerInterface $configurationManager)
     {
         $this->configurationManager = $configurationManager;
-    }
-
-    /**
-     * /**
-     * Returns the Edit Icon with link
-     *
-     * @param string $table Table name
-     * @param array $row Data row
-     * @return string html output
-     */
-    protected function getEditIcon($table, array $row)
-    {
-
-        $params .= '&edit[' . $table . '][' . $row['uid'] . ']=edit';
-        $title = LocalizationUtility::translate('be.editEvent', 'slub_events',
-                $arguments = null) . ' ' . $row['uid'] . ': ' . $row['title'];
-        $icon = '<a href="#" onclick="' . htmlspecialchars(BackendUtility::editOnClick($params, $this->backPath,
-                -1)) . '" title="' . $title . '">' .
-            $this->getSpriteIcon('actions-document-open') .
-            '</a>';
-
-        return $icon;
-    }
-
-    /**
-     * Returns the New Icon with link
-     *
-     * @param string $table Table name
-     * @param array $row Data row
-     * @return string html output
-     */
-    protected function getNewIcon($table, array $row)
-    {
-
-        $params .= '&edit[' . $table . '][' . $row['storagePid'] . ']=new';
-        $title = LocalizationUtility::translate('be.newEvent', 'slub_events', $arguments = null);
-        $icon = '<a href="#" onclick="' . htmlspecialchars(BackendUtility::editOnClick($params, $this->backPath,
-                -1)) . '" title="' . $title . '">' .
-            $this->getSpriteIcon('actions-document-new') .
-            '</a>';
-
-        return $icon;
-    }
-
-    /**
-     * Returns the New Icon with link
-     *
-     * @param string $table Table name
-     * @param array $row Data row
-     * @return string html output
-     */
-    protected function getHideIcon($table, array $row)
-    {
-
-        $doc = $this->getDocInstance();
-        if ($row['hidden']) {
-            $title = LocalizationUtility::translate('be.unhideEvent', 'slub_events', $arguments = null);
-            $params = '&data[' . $table . '][' . $row['uid'] . '][hidden]=0';
-
-            $hideLink = '';
-            $quoteLink = "";
-            if (version_compare(TYPO3_version, '7.6.0', '>=')) {
-                $hideLink = \TYPO3\CMS\Backend\Utility\BackendUtility::getLinkToDataHandlerAction($params,-1);
-            } else {
-                $hideLink = $doc->issueCommand($params,-1);
-                $quoteLink = "'";
-            }
-
-            $icon = '<a href="#" onclick="' . htmlspecialchars('return jumpToUrl(' . $quoteLink . $hideLink . $quoteLink . ');') . '" title="' . $title . '">' .
-                $this->getSpriteIcon('actions-edit-unhide') .
-                '</a>';
-            // Hide
-        } else {
-            $title = LocalizationUtility::translate('be.hideEvent', 'slub_events', $arguments = null);
-            $params = '&data[' . $table . '][' . $row['uid'] . '][hidden]=1';
-
-            $hideLink = '';
-            if (version_compare(TYPO3_version, '7.6.0', '>=')) {
-                $hideLink = \TYPO3\CMS\Backend\Utility\BackendUtility::getLinkToDataHandlerAction($params,-1);
-            } else {
-                $hideLink = $doc->issueCommand($params,-1);
-                $quoteLink = "'";
-            }
-
-            $icon = '<a href="#" onclick="' . htmlspecialchars('return jumpToUrl(' . $quoteLink . $hideLink . $quoteLink . ');') . '" title="' . $title . '">' .
-                $this->getSpriteIcon('actions-edit-hide') .
-                '</a>';
-        }
-        return $icon;
     }
 
     /**
@@ -207,25 +105,6 @@ class FunctionBarViewHelper extends \TYPO3\CMS\Fluid\ViewHelpers\Be\AbstractBack
         }
     }
 
-    /**
-     * Returns the Datepicker img
-     *
-     * @return string html output
-     */
-    protected function getDatePickerIcon()
-    {
-
-        return $this->getSpriteIcon(
-            'actions-edit-pick-date',
-            array(
-                'style' => 'cursor:pointer;',
-                'id' => 'picker-tceforms-datefield-1',
-                'class' => 't3-icon t3-icon-actions t3-icon-actions-edit t3-icon-edit-pick-date'
-            )
-        );
-
-    }
-
 
     /**
      * Renders a record list as known from the TYPO3 list module
@@ -245,17 +124,19 @@ class FunctionBarViewHelper extends \TYPO3\CMS\Fluid\ViewHelpers\Be\AbstractBack
         }
 
         $frameworkConfiguration = $this->configurationManager->getConfiguration(ConfigurationManagerInterface::CONFIGURATION_TYPE_FRAMEWORK);
-        $row['storagePid'] = $frameworkConfiguration['persistence']['storagePid'];
+        $storagePid = $frameworkConfiguration['persistence']['storagePid'];
+
+        $iconHelper = $this->objectManager->get(\Slub\SlubEvents\Helper\IconsHelper::class);
 
         switch ($icon) {
             case 'new':
-                $content = $this->getNewIcon('tx_slubevents_domain_model_event', $row);
+                $content = $iconHelper->getNewIcon('tx_slubevents_domain_model_event', $storagePid);
                 break;
             case 'edit':
-                $content = $this->getEditIcon('tx_slubevents_domain_model_event', $row);
+                $content = $iconHelper->getEditIcon('tx_slubevents_domain_model_event', $row);
                 break;
             case 'hide':
-                $content = $this->getHideIcon('tx_slubevents_domain_model_event', $row);
+                $content = $iconHelper->getHideIcon('tx_slubevents_domain_model_event', $row['uid'], $row['hidden']);
                 break;
             case 'geniusbar':
                 $content = $this->getGeniusBarIcon($event);
@@ -264,7 +145,7 @@ class FunctionBarViewHelper extends \TYPO3\CMS\Fluid\ViewHelpers\Be\AbstractBack
                 $content = $this->getRecurringIcon($event);
                 break;
             case 'datepicker':
-                $content = $this->getDatePickerIcon();
+                $content = $iconHelper->getDatePickerIcon();
                 break;
         }
 
@@ -272,32 +153,4 @@ class FunctionBarViewHelper extends \TYPO3\CMS\Fluid\ViewHelpers\Be\AbstractBack
 
     }
 
-    /**
-     * Get the requested Sprite Icon
-     *
-     * compatibility helper for TYPO3 6.2 and 7.7
-     *
-     * @param $iconName
-     * @param $options
-     *
-     * @return full HTML tag
-     */
-
-    private function getSpriteIcon($iconName, $options = [])
-    {
-
-        if (version_compare(TYPO3_version, '7.6.0', '>=')) {
-
-            return $this->iconFactory->getIcon($iconName, Icon::SIZE_SMALL);
-
-        } else {
-
-            return IconUtility::getSpriteIcon($iconName, $options);
-
-        }
-
-    }
-
 }
-
-?>
